chore: migrate to router
