--- apps/api-gateway/tsconfig.app.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
-    "module": "commonjs",
-    "types": ["node", "express"],
-    "esModuleInterop": true
-  },
-  "exclude": ["**/*.spec.ts", "setupTests.ts"],
-  "include": ["**/*.ts"]
-}
