+++ apps/api-users/src/app/modules/user/user.resolver.ts
@@ -2,8 +2,8 @@ import { Resolver, Query, ResolveReference } from '@nestjs/graphql'
 import { CurrentUserId } from '@core/nest/decorators/CurrentUserId'
 import { UseGuards } from '@nestjs/common'
 import { GqlAuthGuard } from '@core/nest/gqlAuthGuard/GqlAuthGuard'
+import { firebaseClient } from '@core/nest/common/firebaseClient'
 import { User } from '../../__generated__/graphql'
-import { firebaseClient } from '../../lib/firebaseClient'
 import { UserService } from './user.service'
 
 @Resolver('User')
