+++ libs/nest/common/src/lib/firebaseClient/index.ts
@@ -1,0 +1,1 @@
+export { firebaseClient } from './firebaseClient'
