+++ libs/nest/decorators/src/lib/CurrentUserId/CurrentUserId.ts
@@ -1,11 +1,11 @@
 import { createParamDecorator } from '@nestjs/common'
 import { get } from 'lodash'
-import { GqlExecutionContext } from '@nestjs/graphql'
 import { AuthenticationError } from 'apollo-server-errors'
+import { firebaseClient } from '@core/nest/common/firebaseClient'
 
-export const CurrentUserId = createParamDecorator((data, context) => {
-  const ctx = GqlExecutionContext.create(context).getContext()
-  const userId = get(ctx.headers, 'user-id')
-  if (userId == null) throw new AuthenticationError('No user id provided')
-  return userId
+export const CurrentUserId = createParamDecorator(async (data, context) => {
+  const token = get(context.headers, 'Authorization')
+  const { uid } = await firebaseClient.auth().verifyIdToken(token)
+  if (uid == null) throw new AuthenticationError('Token is invalid')
+  return uid
 })
